Refactor TCP CLI server to a different file in tcp-cli subfolder
