+++ src/index.ts
@@ -2,3 +2,4 @@ export * from "./Utils";
 export * from "./commands/CommandHandler";
 export * from "./config/Config";
 export * from "./rpc/JsonRpcServer";
+export * from "./tcp-cli/TcpCliServer";
